requirements.txt
blinker==1.6.2
click==8.1.6
Flask==2.3.2
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
requests==2.30.0
Werkzeug==2.3.6
Flask-SQLAlchemy==3.0.5
iso8601==2.0.0
python-dateutil==2.8.2
prometheus-flask-exporter==0.22.4
py-healthcheck==1.10.1
